split cli/utils.ts into focused modules and add shared logger/db helpers
Replace the monolithic cli/utils.ts with four focused modules:
- cli/output.ts: writeOutput, shouldOutputJson, runRead (8 consumers)
- cli/gateway-client.ts: gatewayGet, gatewayPost, toQueryString, asRecord
- cli/logger.ts: shared CLI logger instance + getCliLogger re-export (14 consumers)
- cli/db.ts: initializeDb re-export from memory/db (7 consumers)

Update all 21 command files to import from the new modules instead of
the old utils.ts and ../../util/logger.js / ../../memory/db.js paths.
