+++ assistant/src/cli/commands/autonomy.ts
@@ -4,9 +4,7 @@ import { dirname, join } from "node:path";
 
 import type { Command } from "commander";
 
-import { getCliLogger } from "../../util/logger.js";
-
-const log = getCliLogger("cli");
+import { log } from "../logger.js";
 
 // ---------------------------------------------------------------------------
 // Types & constants
