+++ assistant/src/cli/commands/channels.ts
@@ -1,6 +1,7 @@
 import type { Command } from "commander";
 
-import { gatewayGet, runRead, toQueryString } from "../utils.js";
+import { gatewayGet, toQueryString } from "../gateway-client.js";
+import { runRead } from "../output.js";
 
 export function registerChannelsCommand(program: Command): void {
   const channels = program
