+++ assistant/src/cli/commands/credentials.ts
@@ -14,10 +14,8 @@ import {
   listCredentialMetadata,
   upsertCredentialMetadata,
 } from "../../tools/credentials/metadata-store.js";
-import { getCliLogger } from "../../util/logger.js";
-import { shouldOutputJson, writeOutput } from "../utils.js";
-
-const log = getCliLogger("cli");
+import { log } from "../logger.js";
+import { shouldOutputJson, writeOutput } from "../output.js";
 
 // ---------------------------------------------------------------------------
 // Shared helpers
