+++ assistant/src/cli/commands/notifications.ts
@@ -1,7 +1,6 @@
 import type { Command } from "commander";
 
 import { getDeliverableChannels } from "../../channels/config.js";
-import { initializeDb } from "../../memory/db.js";
 import { emitNotificationSignal } from "../../notifications/emit-signal.js";
 import { listEvents } from "../../notifications/events-store.js";
 import {
@@ -11,10 +10,9 @@ import {
   NOTIFICATION_SOURCE_EVENT_NAMES,
 } from "../../notifications/signal.js";
 import type { NotificationChannel } from "../../notifications/types.js";
-import { getCliLogger } from "../../util/logger.js";
-import { shouldOutputJson, writeOutput } from "../utils.js";
-
-const log = getCliLogger("cli");
+import { initializeDb } from "../db.js";
+import { log } from "../logger.js";
+import { shouldOutputJson, writeOutput } from "../output.js";
 
 // ---------------------------------------------------------------------------
 // Help text builders
