+++ assistant/src/cli/db.ts
@@ -1,0 +1,1 @@
+export { initializeDb } from "../memory/db.js";
